Revert "small changes"
This reverts commit 4f23c3435f00fbf530d92f7bdb05f9cb11151adc.

+++ source/functions_strings.cpp
@@ -31,7 +31,7 @@ namespace purestringfunctions
 
     int is_canonical(std::string & s)
     {
-        for (size_t i = 0; i < s.length(); i++)
+        for (int i = 0; i < s.length(); i++)
         {
             if (s.at(i) < reverse_char(s.at(s.length()-i-1)))
             {
